pyproject.toml
[project]
name = "LymphoMNIST"
version = "0.0.1"
description = "A PyTorch dataset for Lymphocites."
readme = "README.md"
authors = [
    { name = "Khayrul Islam", email = "khayrulbuet13@alum.lehigh.edu" },
]
license = { file="LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch",
    "torchvision",
    "matplotlib",
    "numpy",
    "Pillow",
    "tqdm",
    "requests",
]
requires-python = ">=3.6"

[tool.black]
target-version = ['py38']
include = '\\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | dump
  | .gitignore
  | others
)/
'''

[tool.isort]
profile = "black"